removed unused RMSF file. Moved RMSF to protein_calls
