+++ plugins/protein_calls/include/protein_calls/RMSF.h
@@ -9,7 +9,7 @@
 #include "protein_calls/MolecularDataCall.h"
 
 namespace megamol {
-namespace protein {
+namespace protein_calls {
 
 /**
  * Computes the RMSF for all atoms stored in mol.
